requirements.txt
# Core dependencies
agno>=0.2.0
langfuse>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
openlit  # Required for Agno + Langfuse integration

# Data processing
pandas>=2.0.0
# Note: sqlite3 is built-in with Python

# Web interface
streamlit>=1.28.0

# Optional but recommended
rich>=13.0.0  # For better terminal output

